restructure app component
