+++ src/components/app/App.js
@@ -1,8 +1,8 @@
 import React from "react";
 import "./App.css";
-import StartPage from "./components/StartPage";
+import StartPage from "../StartPage";
 import { Provider } from "react-redux";
-import store from "./store";
+import store from "../../store";
 
 function App() {
   return (
